Fix PosixTZ offset calculation
Part of the problem was that I ported over code from musl libc without
understanding it. It turns out that musl was expecting years to be
encoded as the number of years from 1900! No doubt that came from when
years were handled as two digit numbers.

Another part of the problem was that I misunderstood what the rule
`M3.2.0` meant in PosixTZ. Previously I though it meant "Sunday of the
2nd week of March". I now understand that it means "The 2nd Sunday of
March".

Anyway, I ripped out the code I ported for `year_to_secs` and replaced
it with something less opaque. I also split `isLeapYear` into a separate
function, instead of having `year_to_secs` take a pointer to a boolean.